added deleting and banning people

+++ pages/api/forms/stand/[id].ts
@@ -15,9 +15,11 @@ export default handler
 		return res.status(200).json(form);
 	})
 	.delete(async (req, res) => {
-		if (!req.user.administrator || !req.user)
-			return res.status(401).json({ message: 'You are not authorized to update users.' });
+		if (!req.user.administrator || !req.user || req.user.banned)
+			return res
+				.status(401)
+				.json({ message: 'You are not authorized to delete stand forms.' });
 		const id = String(req.query.id);
-		StandForm.findByIdAndDelete(id);
+		await StandForm.findByIdAndDelete(id);
 		return res.status(200).json({ message: 'Successfully deleted.' });
 	});
